Election 'Managers' couldn't see or manage elections, only creators.
When loading the Elections page. it only shows you the elections you
created, but not the elections you are set as a manager for. (if another
user created an election and set you as an Manager) Also any user was
able to access any election ballot via the /default/start page by
modifying the url, not just the elections they are managers of. You
couldn't modify anything, however you could see ballot/election info you
may not be privy too and also hit the button to start and email the
voters.  Edited the controllers\default.py edit, start, and elections
functions to correct the issues mentioned above.